+++ src/library/main.cpp
@@ -93,6 +93,12 @@ void __attribute__((constructor)) init(void)
                 receiveData(&index, sizeof(int));
                 Checkpoint::setBaseSavestateIndex(index);
                 break;
+            case MSGN_BACKTRACK_SAVESTATE_PATH:
+                backtracksavestatepath = receiveString();
+                break;
+            case MSGN_BACKTRACK_SAVESTATE_INDEX:
+                receiveData(&backtracksavestateindex, sizeof(int));
+                break;
             default:
                 debuglog(LCF_ERROR | LCF_SOCKET, "Unknown socket message ", message);
                 exit(1);