+++ src/program/utils.cpp
@@ -53,7 +53,7 @@ void remove_savestates(Context* context)
 {
     std::string savestateprefix = context->config.savestatedir + '/';
     savestateprefix += context->gamename;
-    for (int i=0; i<=9; i++) {
+    for (int i=0; i<=10; i++) {
         std::string savestatepmpath = savestateprefix + ".state" + std::to_string(i) + ".pm";
         unlink(savestatepmpath.c_str());
         std::string savestatepspath = savestateprefix + ".state" + std::to_string(i) + ".p";