+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 * Improve the support of virtual screen resolution for non-SDL games
 * Prevent Xlib games to switch to fullscreen
 * Implement an auto-restart feature
-* Manually add audio latency for better playback 
+* Manually add audio latency for better playback
+* Add a restart input
+* Add an option to make a savestate each time the thread set has changed
 
 ### Changed
 ### Fixed
@@ -15,6 +17,8 @@
 * Specifying a movie in commandline does not truncate the file anymore
 * Fix a crash when renaming the frame column in input editor
 * Fix behavior when SDL_JoystickOpen is called multiple times
+* Print the correct char for common keys when typing text (non-SDL)
+* Restore file offsets and close files when loading a savestate
 
 ## [1.3.2] - 2018-11-16
 ### Added